+++ CHANGELOG.md
@@ -20,13 +20,14 @@
 - Controls. Annoying autoselect on touchscreens. Now $options.selectOnFocus is disabled by default
 - Combobox controls (Select, Date, Time). Now popup isn't opened if user clears control and gets focus at the same time
 - Combobox controls (Select, Date, Time). Click on disabled item throws console.error
+- Combobox controls (Select, Date, Time). Focus goes to menu-list by keydown 'Tab' in Firefox
 - [SpinElement](src/spinElement.ts). Animation for TwinDualRing doesn't work on Safari 14-
 - [SpinElement](src/spinElement.ts). Wrong render on option `fit` on Safari 14-
 - [PopupElement](src/popup/popupElement.ts). Animation for opacity doesn't work on Safari 14-
 - [PopupElement](src/popup/popupElement.ts). Animation affects on bluring text if user scroll body during the animation
-- [PopupElement](src/popup/popupElement.ts). Fix wrong position during dropdownAnimation and several show/hide at once
+- [PopupElement](src/popup/popupElement.ts). Wrong position during dropdownAnimation and several show/hide at once
 - [PopupElement](src/popup/popupElement.ts). Allow to open/close on double-click
-- [PopupElement](src/popup/popupElement.ts). Fix 2nd $show() has predefined maxSize & position can be different
+- [PopupElement](src/popup/popupElement.ts). 2nd $show() has predefined maxSize & position can be different
 - [PopupElement](src/popup/popupElement.ts). Now popup has max size 100vw & 100vh (by default). Earlier popup could be > 100vw
 - [PopupElement](src/popup/popupElement.ts). No-updates on screensize changes. Now popup position & size changes with device-rotation
 - [PopupElement](src/popup/popupElement.ts). No popup if mousedown>move>up>click on target